# Use ConfigCache instead of symfony/cache