--- a/src/Resources/config/services.yaml
+++ b/src/Resources/config/services.yaml
@@ -10,3 +10,9 @@ services:
     Becklyn\Translations\:
         resource: '../../*'
         exclude: '../../{Resources,BecklynTranslationsBundle.php}'
+
+    Becklyn\Translations\Extractor\TranslationsExtractor:
+        calls:
+            - method: setConfigCacheFactory
+              arguments:
+                  - '@config_cache_factory'