| Q             | A
| ------------- | --------------------------------------------------------------------- |
| BC breaks?    | no                                                                |
| New feature?  | no <!-- don't forget to update CHANGELOG.md -->                   |
| Improvement?  | yes <!-- improves an existing feature, not adding a new one -->    |
| Bug fix?      | no                                                                |
| Deprecations? | no <!-- don't forget to update UPGRADE.md and CHANGELOG.md -->    |
| Docs PR       | — <!-- insert URL here -->                                  |

<!-- describe your changes below -->
The as with the JS router bundle, this one uses the `ConfigCache` instead of just using the `symfony/cache`.

This has the explicit improvement, that we add a new case, which normally heavily improves performance. The four possible cases are:

* symfony mode `debug`, cache empty (normal debug)
* symfony mode `debug`, cache filled (was not possible previously, as we never used the cache in `debug`)
* symfony mode `prod`, cache empty (not interesting, as practically never happens. Should stay as-is though)
* symfony mode `prod`, cache filled (normal prod)

Here is the comparison:

|  | dev (cache empty) | dev (cache filled) | prod (cache filled) |
| --- | ---- | ---- | --- |
| before | `680ms` | (`680ms`) \* | `0.2ms` |
| after | `690ms` | `8ms` | `0.04ms` |

\*) no special case, as we never used the cache in debug, so the "cache was always empty"